feat: 최종본

See merge request sw_track/class_02_seoul/web_project/team17/mainproject!37

+++ src/routers/profile-router.js
@@ -1,14 +1,12 @@
 import { Router } from "express";
 import is from "@sindresorhus/is";
-// 폴더에서 import하면, 자동으로 폴더의 index.js에서 가져옴
 import { loginRequired } from "../middlewares";
 import { userService, productService } from "../services";
-import { productModel, userModel } from "../db";
+import bcrypt from "bcrypt";
 
 const profileRouter = Router();
 
 //GET: 사용자 정보 가져오기
-//loginRequired 를 통해 userId를 받아온 후, user의 정보를 가져옴
 profileRouter.get("/myProfile", loginRequired, async function (req, res) {
     const userId = req.currentUserId;
     const user = await userService.getUser(userId);
@@ -27,16 +25,12 @@ profileRouter.patch("/edit", loginRequired, async function (req, res, next) {
             );
         }
 
-        // params로부터 id를 가져옴
         const userId = req.currentUserId;
-        // body data 로부터 업데이트할 사용자 정보를 추출함.
         const fullName = req.body.fullName;
         const password = req.body.password;
         const address = req.body.address;
         const phoneNumber = req.body.phoneNumber;
         const role = req.role;
-        console.log(role);
-        // body data로부터, 확인용으로 사용할 현재 비밀번호를 추출함.
         const currentPassword = req.body.currentPassword;
 
         // currentPassword 없을 시, 진행 불가
@@ -45,7 +39,6 @@ profileRouter.patch("/edit", loginRequired, async function (req, res, next) {
         }
 
         const userInfoRequired = { userId, currentPassword };
-
         // 위 데이터가 undefined가 아니라면, 즉, 프론트에서 업데이트를 위해
         // 보내주었다면, 업데이트용 객체에 삽입함.
         const toUpdate = {
@@ -55,6 +48,8 @@ profileRouter.patch("/edit", loginRequired, async function (req, res, next) {
             ...(phoneNumber && { phoneNumber }),
             ...(role && { role }),
         };
+
+        toUpdate.passwordReset = false;
 
         // 사용자 정보를 업데이트함.
 
@@ -73,12 +68,20 @@ profileRouter.patch("/edit", loginRequired, async function (req, res, next) {
 
 profileRouter.delete("/quit", loginRequired, async function (req, res, next) {
     try {
+        //** 현재의 비밀번호가 일치할때 탈퇴 가능하게  */
         const userId = req.currentUserId;
-        const deletedUser = await userService.deleteUser(userId);
-        console.log(deletedUser);
-        res.status(200);
+        const password = req.body.password;
+        const user = await userService.getUser(userId);
+        const isPasswordCorrect = await bcrypt.compare(password, user.password);
+
+        if (!isPasswordCorrect) {
+            throw new Error("비밀번호가 일치하지 않습니다.");
+        } else {
+            const deletedUser = await userService.deleteUser(userId);
+            res.status(200).json(deletedUser);
+        }
     } catch (error) {
-        console.log("탈퇴하기 백엔드에서 에러가 났습니다.");
+        next(error);
     }
 });
